+++ CHANGELOG.md
@@ -6,6 +6,12 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 # Master (will become release 2.10)
 
 
+## Deprecations and removals
+
+- `Dune::Transitional::ReferenceElement` is deprecated and will be removed after
+  Dune 2.10. Use `Dune::Geo::ReferenceElement` directly.
+
+
 # Release 2.9
 
 - The `Geometry` interface was extended by methods `jacobian(local)` and `jacobianInverse(local)`